+++ packages/pc/src/layout/worker/work.ts
@@ -1,9 +1,11 @@
 export default class WebWorker {
   constructor(worker: WebWorker, workerScirptURL: string) {
     const code = worker.toString();
-    const scriptURL = `const workerScirptURL = '${workerScirptURL}';`;
-
-    const blob = new Blob([`${scriptURL}(${code})()`], { type: 'text/javascript' });
+   
+    const blob = new Blob([`importScripts('${workerScirptURL}');(${code})()`], {
+      type: 'text/javascript',
+    });
+    
     return new Worker(URL.createObjectURL(blob));
   }
 }