+++ packages/pc/tests/unit/layout/destroy-layout-spec.ts
@@ -1,5 +1,5 @@
 import G6 from '../../../src';
-import { NodeConfig, EdgeConfig } from '../../../src/types';
+import { NodeConfig, EdgeConfig } from '@antv/g6-core';
 import { mathEqual } from './util';
 
 const div = document.createElement('div');