+++ packages/site/gatsby-browser.js
@@ -1,4 +1,5 @@
 // window.g6 = require('@antv/g6/es'); // import the source for debugging
+// window.g6 = require('@antv/g6/lib'); // import the source for debugging
 window.g6 = require('@antv/g6/dist/g6.min.js'); // import the package for webworker
 window.insertCss = require('insert-css');
 window.Chart = require('@antv/chart-node-g6');