+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+#### 4.2.0
+- fix: combo edge link position problem;
+- fix: activate-relations with combo and combo edges problem;
+- feat: support config TimeBar handler, background, foreground, tick label, tick line style;
+
+#### 4.1.16
+- fix: webworker in dist;
+
 #### 4.1.15
 - fix: cubic-x problem, closes: #2698;